Move generate_table from cranelift-codegen-shared to cranelift-codegen-meta
