+++ cranelift/codegen/meta/src/gen_settings.rs
@@ -1,7 +1,8 @@
 //! Generate the ISA-specific settings.
 use std::collections::HashMap;
 
-use cranelift_codegen_shared::constant_hash::{generate_table, simple_hash};
+use crate::constant_hash::generate_table;
+use cranelift_codegen_shared::constant_hash::simple_hash;
 
 use crate::cdsl::camel_case;
 use crate::cdsl::settings::{
